requirements.txt
pygame==2.5.2
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
numpy==1.24.3
noise==1.2.2
Pillow==10.0.0
moderngl==5.8.2
glfw==2.6.2
